kafka_api_producer/requirements.txt
requests
confluent-kafka==2.8.0
python-dotenv
colorlog==6.9.0